#43: Remove the @types dir
