--- src/@types/ofpData.ts
@@ -1,8 +1,0 @@
-interface OfpOutcome {
-  team: string
-  pointsPercent: number
-  rank?: number
-  won?: boolean
-}
-
-export type OfpData = OfpOutcome[][]
