+++ src/main.ts
@@ -6,15 +6,13 @@ import flags from '@config/flags'
 import TEST_ODDS_DATA_JSON from '@data/testOddsData.json'
 import TEST_OFP_DATA_JSON from '@data/testOfficeFootballPoolData.json'
 import TEST_SCHEDULE_DATA_JSON from '@data/testScheduleData.json'
-import { getOddsData } from '@service/odds'
+import { getOddsData, OddsData } from '@service/odds'
 import { getDailyFirstGames, getNflGamesThisWeek, ScheduleData } from '@service/schedule'
 import { generateOutcomes, mergeOfpAndOddsData } from '@utils/game'
 import { displayPicks } from '@utils/display'
 import { simulateWeek } from '@utils/simulator'
-import { getOfpData } from '@webscraper/getData'
+import { getOfpData, OfpData } from '@webscraper/getData'
 import { makePicks } from '@webscraper/makePicks'
-import { OddsData } from './@types/oddsData'
-import { OfpData } from './@types/ofpData'
 
 const TEST_OFP_DATA = TEST_OFP_DATA_JSON as unknown as OfpData
 const TEST_ODDS_DATA = TEST_ODDS_DATA_JSON as unknown as OddsData
