+++ src/picks/bestSeasonPicks.ts
@@ -1,7 +1,6 @@
-import { GameData, ResultPoints } from '../@types/gameData'
+import { getFixedGames, GameData, ResultPoints } from '@utils/game'
 import { BestPicks, getBestPicks } from '@picks/bestPicks'
 import { AVG_STANDARD_DEVIATION, NET_TARGET } from '@config/constants'
-import { getFixedGames } from '@utils/game'
 import { zScoreToProb } from '@utils/stats'
 
 export const getBestSeasonPicks = (games: GameData[], outcomes: ResultPoints[][][]): [BestPicks, number] => {
