+++ src/picks/bestWeeklyPicks.ts
@@ -1,9 +1,8 @@
 import cliProgress from 'cli-progress'
 
-import { GameData, ResultPoints } from '../@types/gameData'
+import { getFixedGames, GameData, ResultPoints } from '@utils/game'
 import { BestPicks, GameToPick, getBestPicks } from '@picks/bestPicks'
 import { AVG_STANDARD_DEVIATION, MAX_RANK, NET_TARGET } from '@config/constants'
-import { getFixedGames } from '@utils/game'
 import { chooseK, possibleScenarios, zScoreToProb } from '@utils/stats'
 
 interface BestProb {
