+++ src/utils/simulator.ts
@@ -1,4 +1,4 @@
-import { GameData, ResultPoints } from '../@types/gameData'
+import { GameData, ResultPoints } from '@utils/game'
 import { MAX_RANK } from '@config/constants'
 
 export const simulateWeek = (games: GameData[], outcomes: ResultPoints[][][], picks: Array<[number, number]>, numSimulations = 100_000): void => {
